pyproject.toml
[project]
name = "repl-nix-projectrevive"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=2.0.0",
    "flask-login>=0.6.2",
    "flask-sqlalchemy>=3.1.1",
    "oauthlib>=3.2.2",
    "psycopg2-binary>=2.9.10",
    "flask-migrate>=4.0.7",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=1.4.0",
    "flask-cors>=5.0.0",
    "flask-babel>=3.1.0",
    "werkzeug>=3.0.6",
    "itsdangerous>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask-mail>=0.10.0",
    "openai>=1.0.0",
    "pandas",
    "tenacity>=8.2.3",
    "pinecone-client>=2.2.4",
    "pymupdf==1.23.8",
    "fitz>=0.0.1.dev2",
    "pinecone>=5.4.2",
    "pdfplumber>=0.11.4",
    "faiss-cpu>=1.9.0.post1",
    "numpy>=2.1.3",
    "psutil>=6.1.1",
    "aioredis>=2.0.1",
    "cachetools>=5.5.0",
    "async-timeout>=5.0.1",
    "redis>=5.2.1",
    "fakeredis>=2.26.2",
]
